refactor[cm]: improve file structure
