+++ CodeModel/CodeParsers/CSharp/ClassCohesionAnalyzer/Model/ClassPart.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace CodeModel.CodeParsers.CSharp.ClassCohesionAnalyzer
+namespace CodeModel.CodeParsers.CSharp.ClassCohesionAnalyzer.Model
 {
     public class ClassPart
     {
